refactor into new source code folder, remove some duplicated code
